+++ NQueen.GUI/ViewModels/InputViewModel.cs
@@ -1,7 +1,5 @@
 ﻿namespace NQueen.GUI.ViewModels;
 
-// Todo: Resolve the following bug:
-// When simulating the second time with the same input, the chessboard is not reset.
 public class InputViewModel(SolutionMode solutionMode)
 {
     public FluentValidation.Results.ValidationResult ValidateBoardSize(string? boardSizeText)