+++ NQueen.Shared/Utils/UiMessages.cs
@@ -1,6 +1,5 @@
 ﻿namespace NQueen.Shared.Utils;
 
-// Todo: Move other UI-related constants here, e.g. from ConsoleApp.Commands.DispatchCommands
 public class UiMessages
 {
     public const string RunAgainPrompt =